requirements.txt
boto3
botocore
mpv
requests
pyyaml
python-mpv